Carry the extension in one file, and package it into the archives
The compiler is one C file; the extension is now one JavaScript file. The
manifest, the grammar, the language configuration, the snippets and the two
packaging files ride in ada83-extension.js as line comments, and `make vsix`
splits them back out into the layout VS Code installs.

Line comments are the only container that works here. A block comment ends at
the */ inside the grammar's operator character class, and a template literal
interpolates at the ${ in every snippet, so both would be broken by their own
payload; a line comment cannot be ended by anything a line can hold.

A .vsix is a zip with the manifest and sources under extension/ and the two
packaging files at the root, so building one needs no tooling beyond zip on
one side and Compress-Archive on the other. `make vsix` and `make.bat vsix`
build it alone; `package` puts it in the archive beside the compiler, so one
download carries the compiler, the runtime and the editor support.

Verified by installing the result: "Extension 'ada83.vsix' was successfully
installed", and every split-out file re-parses as the JSON or XML it claims
to be.
