--- vscode/language-configuration.json
@@ -1,20 +1,0 @@
-{
-  "comments": { "lineComment": "--" },
-  "brackets": [["(", ")"]],
-  "autoClosingPairs": [
-    { "open": "(", "close": ")" },
-    { "open": "\"", "close": "\"", "notIn": ["string", "comment"] }
-  ],
-  "surroundingPairs": [["(", ")"], ["\"", "\""]],
-  "wordPattern": "[A-Za-z][A-Za-z0-9_]*",
-  "onEnterRules": [
-    {
-      "beforeText": "^\\s*--.*$",
-      "action": { "indent": "none", "appendText": "-- " }
-    }
-  ],
-  "indentationRules": {
-    "increaseIndentPattern": "(?i)^\\s*(.*\\b(is|then|loop|begin|record|declare|select|do|else)\\s*|.*\\b(procedure|function|package|task)\\b.*)$",
-    "decreaseIndentPattern": "(?i)^\\s*\\b(end|elsif|else|exception|when|begin|private)\\b"
-  }
-}
