--- vscode/package.json
@@ -1,88 +1,0 @@
-{
-  "name": "ada83",
-  "displayName": "Ada 83",
-  "description": "Ada 83 (ANSI/MIL-STD-1815A) support, with diagnostics from the ada83 compiler itself.",
-  "version": "1.0.0",
-  "publisher": "ada83",
-  "license": "SEE LICENSE IN ../readme.md",
-  "engines": {
-    "vscode": "^1.75.0"
-  },
-  "categories": [
-    "Programming Languages",
-    "Linters"
-  ],
-  "activationEvents": [
-    "onLanguage:ada83"
-  ],
-  "main": "./extension.js",
-  "contributes": {
-    "languages": [
-      {
-        "id": "ada83",
-        "aliases": [
-          "Ada 83",
-          "ada"
-        ],
-        "extensions": [
-          ".ada",
-          ".adb",
-          ".ads"
-        ],
-        "configuration": "./language-configuration.json"
-      }
-    ],
-    "grammars": [
-      {
-        "language": "ada83",
-        "scopeName": "source.ada",
-        "path": "./syntaxes/ada83.tmLanguage.json"
-      }
-    ],
-    "snippets": [
-      {
-        "language": "ada83",
-        "path": "./snippets.json"
-      }
-    ],
-    "configuration": {
-      "title": "Ada 83",
-      "properties": {
-        "ada83.compilerPath": {
-          "type": "string",
-          "default": "ada83",
-          "description": "Path to the ada83 compiler. It is run as `ada83 --lsp`, and looks for ada83-runtime.ada beside itself."
-        },
-        "ada83.enable": {
-          "type": "boolean",
-          "default": true,
-          "description": "Report diagnostics from the compiler as you type."
-        }
-      }
-    },
-    "problemMatchers": [
-      {
-        "name": "ada83",
-        "owner": "ada83",
-        "source": "ada83",
-        "fileLocation": [
-          "autoDetect",
-          "${workspaceFolder}"
-        ],
-        "pattern": {
-          "regexp": "^(\\S.*?):(\\d+):(\\d+):\\s+(error|warning|note|internal error):\\s+(.*)$",
-          "file": 1,
-          "line": 2,
-          "column": 3,
-          "severity": 4,
-          "message": 5
-        }
-      }
-    ],
-    "taskDefinitions": [
-      {
-        "type": "ada83"
-      }
-    ]
-  }
-}
