--- vscode/snippets.json
@@ -1,66 +1,0 @@
-{
-  "procedure body": {
-    "prefix": "procedure",
-    "body": ["procedure ${1:Name} is", "begin", "   $0", "end ${1:Name};"],
-    "description": "A procedure body"
-  },
-  "function body": {
-    "prefix": "function",
-    "body": ["function ${1:Name} return ${2:Type} is", "begin", "   $0",
-             "end ${1:Name};"],
-    "description": "A function body"
-  },
-  "package specification": {
-    "prefix": "package",
-    "body": ["package ${1:Name} is", "   $0", "end ${1:Name};"],
-    "description": "A package specification"
-  },
-  "package body": {
-    "prefix": "packagebody",
-    "body": ["package body ${1:Name} is", "   $0", "end ${1:Name};"],
-    "description": "A package body"
-  },
-  "for loop": {
-    "prefix": "for",
-    "body": ["for ${1:I} in ${2:1 .. 10} loop", "   $0", "end loop;"],
-    "description": "A for loop"
-  },
-  "while loop": {
-    "prefix": "while",
-    "body": ["while ${1:Condition} loop", "   $0", "end loop;"],
-    "description": "A while loop"
-  },
-  "if statement": {
-    "prefix": "if",
-    "body": ["if ${1:Condition} then", "   $0", "end if;"],
-    "description": "An if statement"
-  },
-  "case statement": {
-    "prefix": "case",
-    "body": ["case ${1:Expression} is", "   when ${2:Choice} =>", "      $0",
-             "   when others =>", "      null;", "end case;"],
-    "description": "A case statement"
-  },
-  "declare block": {
-    "prefix": "declare",
-    "body": ["declare", "   ${1:Name} : ${2:Type};", "begin", "   $0", "end;"],
-    "description": "A block statement with declarations"
-  },
-  "record type": {
-    "prefix": "record",
-    "body": ["type ${1:Name} is", "   record", "      ${2:Field} : ${3:Type};",
-             "   end record;"],
-    "description": "A record type declaration"
-  },
-  "exception handler": {
-    "prefix": "exception",
-    "body": ["exception", "   when ${1:Constraint_Error} =>", "      $0"],
-    "description": "An exception handler"
-  },
-  "hello world": {
-    "prefix": "hello",
-    "body": ["with Text_IO; use Text_IO;", "procedure ${1:Hello} is", "begin",
-             "   Put_Line (\"${2:Hello, Ada 83!}\");", "end ${1:Hello};"],
-    "description": "A complete program"
-  }
-}
