--- vscode/syntaxes/ada83.tmLanguage.json
@@ -1,188 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/martinring/tmlanguage/master/tmlanguage.json",
-  "name": "Ada 83",
-  "scopeName": "source.ada",
-  "patterns": [
-    {
-      "include": "#comment"
-    },
-    {
-      "include": "#string"
-    },
-    {
-      "include": "#character"
-    },
-    {
-      "include": "#number"
-    },
-    {
-      "include": "#declaration"
-    },
-    {
-      "include": "#attribute"
-    },
-    {
-      "include": "#predefined"
-    },
-    {
-      "include": "#literal"
-    },
-    {
-      "include": "#keyword"
-    },
-    {
-      "include": "#parameter"
-    },
-    {
-      "include": "#operator"
-    }
-  ],
-  "repository": {
-    "comment": {
-      "name": "comment.line.double-dash.ada",
-      "match": "--.*$"
-    },
-    "string": {
-      "name": "string.quoted.double.ada",
-      "begin": "\"",
-      "end": "\"",
-      "patterns": [
-        {
-          "name": "constant.character.escape.ada",
-          "match": "\"\""
-        }
-      ]
-    },
-    "character": {
-      "name": "string.quoted.single.ada",
-      "match": "'(?:[^']|'')'"
-    },
-    "number": {
-      "patterns": [
-        {
-          "name": "constant.numeric.based.ada",
-          "match": "\\b\\d+#[0-9A-Fa-f_]+#(?:[Ee][+-]?\\d+)?"
-        },
-        {
-          "name": "constant.numeric.decimal.ada",
-          "match": "\\b\\d[\\d_]*(?:\\.[\\d_]+)?(?:[Ee][+-]?\\d+)?\\b"
-        }
-      ]
-    },
-    "keyword": {
-      "patterns": [
-        {
-          "name": "keyword.control.ada",
-          "match": "(?i)\\b(abort|accept|access|all|array|at|begin|body|case|constant|declare|delay|delta|digits|do|else|elsif|end|entry|exception|exit|for|function|generic|goto|if|in|is|limited|loop|new|null|of|others|out|package|pragma|private|procedure|raise|range|record|renames|return|reverse|select|separate|subtype|task|terminate|then|type|use|when|while|with)\\b"
-        },
-        {
-          "name": "keyword.operator.word.ada",
-          "match": "(?i)\\b(abs|and|mod|not|or|rem|xor)\\b"
-        }
-      ]
-    },
-    "attribute": {
-      "name": "support.other.attribute.ada",
-      "match": "'[A-Za-z][A-Za-z0-9_]*"
-    },
-    "operator": {
-      "name": "keyword.operator.ada",
-      "match": "(:=|=>|\\.\\.|\\*\\*|/=|>=|<=|<<|>>|<>|[+\\-*/&<>=|])"
-    },
-    "declaration": {
-      "patterns": [
-        {
-          "match": "(?i)\\b(procedure|function|entry)\\s+([A-Za-z][A-Za-z0-9_]*)",
-          "captures": {
-            "1": {
-              "name": "keyword.control.ada"
-            },
-            "2": {
-              "name": "entity.name.function.ada"
-            }
-          }
-        },
-        {
-          "match": "(?i)\\b(package)\\s+(body\\s+)?([A-Za-z][A-Za-z0-9_]*)",
-          "captures": {
-            "1": {
-              "name": "keyword.control.ada"
-            },
-            "2": {
-              "name": "keyword.control.ada"
-            },
-            "3": {
-              "name": "entity.name.namespace.ada"
-            }
-          }
-        },
-        {
-          "match": "(?i)\\b(task)\\s+(body\\s+|type\\s+)?([A-Za-z][A-Za-z0-9_]*)",
-          "captures": {
-            "1": {
-              "name": "keyword.control.ada"
-            },
-            "2": {
-              "name": "keyword.control.ada"
-            },
-            "3": {
-              "name": "entity.name.class.ada"
-            }
-          }
-        },
-        {
-          "match": "(?i)\\b(type|subtype)\\s+([A-Za-z][A-Za-z0-9_]*)",
-          "captures": {
-            "1": {
-              "name": "keyword.control.ada"
-            },
-            "2": {
-              "name": "entity.name.type.ada"
-            }
-          }
-        },
-        {
-          "match": "(?i)\\b(generic|pragma)\\s+([A-Za-z][A-Za-z0-9_]*)?",
-          "captures": {
-            "1": {
-              "name": "keyword.control.ada"
-            },
-            "2": {
-              "name": "entity.name.tag.ada"
-            }
-          }
-        }
-      ]
-    },
-    "predefined": {
-      "name": "support.type.ada",
-      "match": "(?i)\\b(Integer|Natural|Positive|Float|Boolean|Character|String|Duration|Short_Integer|Long_Integer|Short_Float|Long_Float|System|Standard)\\b"
-    },
-    "literal": {
-      "patterns": [
-        {
-          "name": "constant.language.boolean.ada",
-          "match": "(?i)\\b(True|False)\\b"
-        },
-        {
-          "name": "support.class.exception.ada",
-          "match": "(?i)\\b(Constraint_Error|Numeric_Error|Program_Error|Storage_Error|Tasking_Error)\\b"
-        }
-      ]
-    },
-    "parameter": {
-      "match": "(?i)\\b([A-Za-z][A-Za-z0-9_]*)\\s*(:)\\s*((?:in\\s+out|in|out)\\b)?",
-      "captures": {
-        "1": {
-          "name": "variable.parameter.ada"
-        },
-        "2": {
-          "name": "punctuation.separator.ada"
-        },
-        "3": {
-          "name": "keyword.control.ada"
-        }
-      }
-    }
-  }
-}
